Fix #5 by using params in URL from Search.js to server.js
Current Version : v1.1

I came up with a way to pass data as a parameter through an address in "Search.js".
I was wondering if this could be used in "app.use" method used in the "server.js",
so I applied it. I ended up making improvements with this one.