Fix #7 by using "dot-env-webpack"
Current Version : v1.3

- Why the error occured?
At the Error messages, I realized that there are no fs, os, path modules.
So, I decided to study what they really are. Before that, I looked up what dotenv was.

dotenv is a zero-dependency module that loads variables declared in the .env file into process.env.
This dotdenv finds the absolute path of the .env file through the os and path modules,
and physically reads the .env file with the fs module and stores it in process.env in key-value format.
I was able to figure out the problem here.

The os, path, and fs modules are one of the modules provided by Node.js by default.
As such, dotenv is a package intended to run in the default environment of Node.js,
so it can be said to be an independent module.
When I call dotenv in client.js in src, it is executed in a location where the module cannot be loaded.

- How I fix it?
I added the "dotenv-webpack" package for the "webpack.config.js". It can import the dotenv module in
webpack environment. After that, I added the dotenv func in plugin of "module.exports".