Update the search page using absolute position
Current Version : v1.4

- What I make
I made an asymmetric and freewheeling search page img mainly using “position absolute”.
The most important technique used here is that various props are sent from
the component to the styled emotion css.

- The reason to do this
Through this, it is convenient to dynamically transmit css according to the state
change at the component stage. Also, because this process happens with re-rendering,
changes can be better adapted.