Fix #8 atom for managing the Fbase Login state
Current Version : v1.4

- What I make
By managing the state globally using recoil, various components that need to use
the Firebase login state can be re-rendered at the same time.

- The reason to do this
Previously, Firebase login state was managed using authService.onAuthStateChanged
for each component. However, at this time, an error occurs because the time when the
data of firebase's authservice is reloaded and the time when the react component is
re-rendered are different. Therefore, recoil manages the login state globally to prevent this.

- Principals of them
It allows you to manage state globally using recoil in your code.
Here, the smallest unit of state that can be shared between components is called an atom.
It is a concept very similar to state, but other components can share the state.