Deployt my app using gh=pages with virtual machine
Current Version : v1.4

- What I make
Now let's try to deply the app I made. The process will be done with gh-pages.
Also, I was receiving the interpark API as my CP's server,
but I use Google Cloud Platform to change it. Rent a Virtual Machine,
launch a server there, and get it as an external IP.

Replaced all [localhost:3085] with 34.83.61.147:3085.

- The reason to do this
Github Pages is a service that hosts static pages. I used this service because
it was extremely convenient because it directly hosts the source code of the uploaded GitHub repository.

And Google Cloud Platform is a service that Google provides a server for a fee.
This allows you to use the Interpark API.