Fix #9 using basename and public path
Current Version : v1.4

- How I fixed?
If the load resource failed, it means that the root setting in app.js is wrong.
So it goes into webpack.config.js , which configures this route. Here, when
I changed the public path from ‘/dist/’ to ‘/Book_Helper/’, the app.js file is sent
to the client side smoothly.

Existing code is to say the main page is https://keinn51.github.io/ !
Therefore, you have to put basename=”Bok_Helper” to get the main page as
https://keinn51.github.io/Book_Helper/.