Start Typescript, adding node packages
Current Version : v2.0

- What I make
Pre-configure to use typescript. Replace jsconfig.json with tsconfig.json.
After that, install the following packages.

```jsx
npm i typescript @types/react @types/react-dom ts-node
npm i -D @babel/preset-typescript @types/webpack
```

Typescript is not a language that runs on the web, but a language
that is converted to js through babel. Therefore, in order to read the
code in node.js, you need to download the above packages.

**typescript** → This is an integrated package for typescript.
Includes over 50 packages for typescript including @types/node
and @types/browserify, @types/convert-source-map and more.

**@types/react @types/react-dom @types/webpack**→ It is not included
in the typescript basic package. To use typescript in react, you need to download it separately.

 **ts-node** → This is a pacakage that allows you to run ts in the node runtime environment.

 - The reason to do this
The strong type system used in systematic and refined languages such as C#
and Java can provide high readability and code quality, and it is easier to
catch fatal errors because errors occur in the compilation environment rather than runtime.

On the other hand, JavaScript is a **dynamic programming language** without a
type system, and JavaScript variables can have values of several types, such as
strings, numbers, and booleans. This can be expressed as weakly typed language
and is relatively flexible. While providing an environment that can be developed easily,