Fix #10 paste the index.html to 404.html
Current Version : v2.0

- What problem
After deploying to github pages, there was a problem in which a 404
page was displayed when refreshed on a routed page.

- The reason to the error
github pages do not support SPA. So, even if the URL is refreshed at https://[domain],
the page looks good, but when reloading at https://[domain]/main, the GitHub page server
does not know /main, so a 404 error and 404.html page are returned.

- How to fix
When using GitHub Pages, it's a shortcut, but I use 404.html a lot.
It's because it provides it by itself. In CI, you just need to add a script
that copies index.html to 404.html ;)