Use redux for global state
Current Version : v2.1

- What I make
Replaced the existing state management library from recoil to redux.
I made a folder for redux and added a dispatch file for login of firebase.

- The reason to do this
I checked the number of downloads of the recently used state management
library on the npm trends site. We looked at redux, mobx, and recoil. As a result,
redux boasted a huge number of downloads, while recoil received almost
no downloads. So, as I have never used redux before, I thought I should
definitely use this library, and I tried to apply it to my project.

- Principals of them
To put it simply in words, after receiving an event from the user in the
view and bringing it to the dispatcher, the dispatcher sends it to the store
as an action. Inside the store, there is a function called reducer, which reads
an object called action and updates the state. This updates the state (not a change!)
and sends it back to the view. Then the user will see the UI changed by the updated
state.