Refactor the AuthSocialLogin with atomic design
Current Version : v2.1

- The Problem of the project now
A detailed directory analysis revealed several problems.

 1. The directory name of atom is wrong. I wish it would have been better
if it was called global. Also, in this directory, I thought it would be
better to send the list object to util and replace the recoil with redux.

 2. A directory called hooks is rarely used. We do not make our own hooks,
we only use the hook called useInput that was taught in class. You should
either create more hooks or delete this directory altogether. Or maybe there
is a way to combine it with utils.

 3. There are too many and scattered directories in the component directory.
There are three directories related to Auth and three directories related to
slider, so there is no sense of structure at all.

- Adjust the atomic design pattern
The biggest advantage of atomic design is that it is easy to reuse components.
Since the code is cut for each atom, it is easy to take it from another component
and use it again. I thought it was a necessary part due to the nature of my project
with low component reusability.

However, the downside is that it can be difficult to use media queries because
it is difficult to share css between atoms, molecules and organisms in atomic design.
This indicates that flex and grid are used and should be used.